modules/compute/aws-batch/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.190
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.164
    # via aws-cdk-lib
aws-cdk-aws-batch-alpha==2.70.0a0
    # via -r requirements.in
aws-cdk-lib==2.70.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-batch-alpha
    #   cdk-nag
boto3==1.21.46
    # via -r requirements.in
botocore==1.24.46
    # via
    #   boto3
    #   s3transfer
cattrs==22.2.0
    # via jsii
cdk-nag==2.12.29
    # via -r requirements.in
constructs==10.0.91
    # via
    #   -r requirements.in
    #   aws-cdk-aws-batch-alpha
    #   aws-cdk-lib
    #   cdk-nag
importlib-resources==5.12.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.83.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-batch-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-batch-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-batch-alpha
    #   aws-cdk-lib
    #   jsii
typing-extensions==4.6.3
    # via jsii
urllib3==1.26.18
    # via botocore
